move test to test_utils.py
