+++ tests/ignite/test_utils.py
@@ -2,9 +2,11 @@ import os
 import logging
 import pytest
 import torch
+import torch.distributed as dist
 
 from collections import namedtuple
-from ignite.utils import convert_tensor, to_onehot, setup_logger
+from ignite.utils import convert_tensor, to_onehot, setup_logger, one_rank_only
+from ignite.engine import Engine, Events
 
 
 def test_convert_tensor():
@@ -113,3 +115,79 @@ def test_setup_logger(capsys, dirname):
     for source in [err, data]:
         assert "trainer INFO: Engine run starting with max_epochs=5." in source[0]
         assert "evaluator INFO: Engine run starting with max_epochs=1." in source[2]
+
+
+def _test_distrib_one_rank_only():
+
+    def _test(barrier):
+        # last rank
+        rank = dist.get_world_size() - 1
+
+        value = torch.tensor(0)
+
+        @one_rank_only(rank=rank, barrier=barrier)
+        def initialize():
+            value.data = torch.tensor(100)
+
+        initialize()
+
+        value_list = [torch.tensor(0) for _ in range(dist.get_world_size())]
+
+        dist.all_gather(tensor=value, tensor_list=value_list)
+
+        for r in range(dist.get_world_size()):
+            if r == rank:
+                assert value_list[r].item() == 100
+            else:
+                assert value_list[r].item() == 0
+
+    _test(barrier=True)
+    _test(barrier=False)
+
+
+def _test_distrib_one_rank_only_with_engine():
+
+    def _test(barrier):
+        engine = Engine(lambda e, b: b)
+
+        batch_sum = torch.tensor(0)
+
+        @engine.on(Events.ITERATION_COMPLETED)
+        @one_rank_only(barrier=barrier)  # ie rank == 0
+        def _(_):
+            batch_sum.data += torch.tensor(engine.state.batch)
+
+        engine.run([1, 2, 3], max_epochs=2)
+
+        value_list = [torch.tensor(0) for _ in range(dist.get_world_size())]
+
+        dist.all_gather(tensor=batch_sum, tensor_list=value_list)
+
+        for r in range(dist.get_world_size()):
+            if r == 0:
+                assert value_list[r].item() == 12
+            else:
+                assert value_list[r].item() == 0
+
+    _test(barrier=True)
+    _test(barrier=False)
+
+
+@pytest.mark.distributed
+def test_distrib_cpu(distributed_context_single_node_gloo):
+    _test_distrib_one_rank_only()
+    _test_distrib_one_rank_only_with_engine()
+
+
+@pytest.mark.multinode_distributed
+@pytest.mark.skipif("MULTINODE_DISTRIB" not in os.environ, reason="Skip if not multi-node distributed")
+def test_multinode_distrib_cpu(distributed_context_multi_node_gloo):
+    _test_distrib_one_rank_only()
+    _test_distrib_one_rank_only_with_engine()
+
+
+@pytest.mark.multinode_distributed
+@pytest.mark.skipif("GPU_MULTINODE_DISTRIB" not in os.environ, reason="Skip if not multi-node distributed")
+def test_multinode_distrib_gpu(distributed_context_multi_node_nccl):
+    _test_distrib_one_rank_only()
+    _test_distrib_one_rank_only_with_engine()
